+++ lib/Extended_ast.mli
@@ -27,7 +27,7 @@ type 'a t =
   | Repl_file : repl_file t
 
 module Parse : sig
-  val ast : 'a t -> Lexing.lexbuf -> 'a
+  val ast : 'a t -> preserve_beginend:bool -> Lexing.lexbuf -> 'a
 end
 
 val equal_core_type : core_type -> core_type -> bool