+++ lib/Translation_unit.ml
@@ -331,10 +331,11 @@ let format (type a b) (fg : a Extended_ast.t) (std_fg : b Std_ast.t)
         |> List.filter_map ~f:(fun (s, f_opt) ->
                Option.map f_opt ~f:(fun f -> (s, String.sexp_of_t f)) )
       in
-      ( match
-          parse Extended_ast.Parse.ast ~disable_w50:true fg conf
-            ~source:fmted
-        with
+      let preserve_beginend =
+        Poly.(conf.fmt_opts.exp_grouping = `Preserve)
+      in
+      let parse_ast = Extended_ast.Parse.ast ~preserve_beginend in
+      ( match parse parse_ast ~disable_w50:true fg conf ~source:fmted with
       | exception Sys_error msg -> Error (Error.User_error msg)
       | exception exn -> internal_error (`Cannot_parse exn) (exn_args ())
       | t_new -> Ok t_new )
@@ -478,8 +479,9 @@ let normalize_eol ~strlocs ~line_endings s =
 let parse_and_format (type a b) (fg : a Extended_ast.t)
     (std_fg : b Std_ast.t) ?output_file ~input_name ~source (conf : Conf.t) =
   Location.input_name := input_name ;
-  parse_result Extended_ast.Parse.ast ~disable_w50:true fg conf ~source
-    ~input_name
+  let preserve_beginend = Poly.(conf.fmt_opts.exp_grouping = `Preserve) in
+  let parse_ast = Extended_ast.Parse.ast ~preserve_beginend in
+  parse_result parse_ast ~disable_w50:true fg conf ~source ~input_name
   >>= fun parsed ->
   parse_result Std_ast.Parse.ast std_fg conf ~source ~input_name
   >>= fun std_parsed ->
@@ -514,12 +516,14 @@ let check_range nlines (low, high) =
     Error (Error.User_error (Format.sprintf "Invalid range %i-%i" low high))
 
 let numeric (type a b) (fg : a list Extended_ast.t)
-    (std_fg : b list Std_ast.t) ~input_name ~source ~range conf =
+    (std_fg : b list Std_ast.t) ~input_name ~source ~range (conf : Conf.t) =
   let lines = String.split_lines source in
   let nlines = List.length lines in
   check_range nlines range
   >>| fun () ->
   Location.input_name := input_name ;
+  let preserve_beginend = Poly.(conf.fmt_opts.exp_grouping = `Preserve) in
+  let parse_ast = Extended_ast.Parse.ast ~preserve_beginend in
   let fallback () = Indent.Partial_ast.indent_range ~source ~range in
   let indent_parsed parsed std_parsed ~src ~range =
     let {ast= parsed_ast; _} = parsed in
@@ -527,10 +531,7 @@ let numeric (type a b) (fg : a list Extended_ast.t)
       format fg std_fg ~input_name ~prev_source:src ~parsed ~std_parsed conf
     with
     | Ok (_, fmted_src) -> (
-      match
-        parse_result Extended_ast.Parse.ast fg ~source:fmted_src conf
-          ~input_name
-      with
+      match parse_result parse_ast fg ~source:fmted_src conf ~input_name with
       | Ok {ast= fmted_ast; source= fmted_src; _} ->
           Indent.Valid_ast.indent_range fg ~lines ~range
             ~unformatted:(parsed_ast, src) ~formatted:(fmted_ast, fmted_src)
@@ -538,9 +539,7 @@ let numeric (type a b) (fg : a list Extended_ast.t)
     | Error _ -> fallback ()
   in
   let parse_or_recover ~src =
-    match
-      parse_result Extended_ast.Parse.ast fg conf ~source:src ~input_name
-    with
+    match parse_result parse_ast fg conf ~source:src ~input_name with
     | Ok parsed -> Ok parsed
     | Error _ -> parse_result recover fg conf ~source:src ~input_name
   in